Added filter by all fields

+++ src/main/java/dev/aziz/phonebook/PhonebookApplication.java
@@ -49,37 +49,38 @@ public class PhonebookApplication {
 				System.out.println("Sample users added to the database.");
 			}
 			if (itemRepository.count() == 0) {
-				itemRepository.save(new Item("1", "Screw", 1000, "Units", BigDecimal.valueOf(150)));
-				itemRepository.save(new Item("2", "Nail", 5000, "Units", BigDecimal.valueOf(50)));
-				itemRepository.save(new Item("3", "Hammer", 200, "Units", BigDecimal.valueOf(1200)));
-				itemRepository.save(new Item("4", "Drill Machine", 50, "Units", BigDecimal.valueOf(4500)));
-				itemRepository.save(new Item("5", "Wood Plank", 100, "Units", BigDecimal.valueOf(800)));
-				itemRepository.save(new Item("6", "Cement Bag", 300, "Units", BigDecimal.valueOf(350)));
-				itemRepository.save(new Item("7", "Brick", 5000, "Units", BigDecimal.valueOf(5)));
-				itemRepository.save(new Item("8", "Steel Rod", 100, "Units", BigDecimal.valueOf(1500)));
-				itemRepository.save(new Item("9", "Sand", 2000, "Kilograms", BigDecimal.valueOf(50)));
-				itemRepository.save(new Item("10", "Gravel", 1500, "Kilograms", BigDecimal.valueOf(60)));
-				itemRepository.save(new Item("11", "Paint Bucket", 100, "Units", BigDecimal.valueOf(900)));
-				itemRepository.save(new Item("12", "Ladder", 50, "Units", BigDecimal.valueOf(2500)));
-				itemRepository.save(new Item("13", "Saw", 100, "Units", BigDecimal.valueOf(500)));
-				itemRepository.save(new Item("14", "Shovel", 150, "Units", BigDecimal.valueOf(400)));
-				itemRepository.save(new Item("15", "Trowel", 200, "Units", BigDecimal.valueOf(250)));
-				itemRepository.save(new Item("16", "Plumbing Pipe", 300, "Meters", BigDecimal.valueOf(100)));
-				itemRepository.save(new Item("17", "PVC Glue", 200, "Units", BigDecimal.valueOf(150)));
-				itemRepository.save(new Item("18", "Tile", 400, "Square Meters", BigDecimal.valueOf(300)));
-				itemRepository.save(new Item("19", "Tile Adhesive", 100, "Bags", BigDecimal.valueOf(350)));
-				itemRepository.save(new Item("20", "Glass Sheet", 50, "Units", BigDecimal.valueOf(2000)));
-				itemRepository.save(new Item("21", "Door Handle", 150, "Units", BigDecimal.valueOf(120)));
-				itemRepository.save(new Item("22", "Hinges", 300, "Units", BigDecimal.valueOf(60)));
-				itemRepository.save(new Item("23", "Plaster of Paris", 200, "Bags", BigDecimal.valueOf(250)));
-				itemRepository.save(new Item("24", "Electrical Wire", 500, "Meters", BigDecimal.valueOf(15)));
-				itemRepository.save(new Item("25", "Switchboard", 200, "Units", BigDecimal.valueOf(300)));
-				itemRepository.save(new Item("26", "Breaker Panel", 50, "Units", BigDecimal.valueOf(4000)));
-				itemRepository.save(new Item("27", "Roof Sheet", 100, "Units", BigDecimal.valueOf(2000)));
-				itemRepository.save(new Item("28", "Concrete Mixer", 10, "Units", BigDecimal.valueOf(15000)));
-				itemRepository.save(new Item("29", "Safety Helmet", 200, "Units", BigDecimal.valueOf(500)));
-				itemRepository.save(new Item("30", "Safety Gloves", 300, "Pairs", BigDecimal.valueOf(100)));
-				itemRepository.save(new Item("31", "Welding Rod", 400, "Units", BigDecimal.valueOf(75)));
+				itemRepository.save(new Item("1", "Screw", 1000, "Units", Double.valueOf(150)));
+				itemRepository.save(new Item("2", "Nail", 5000, "Units", Double.valueOf(50)));
+				itemRepository.save(new Item("3", "Hammer", 200, "Units", Double.valueOf(1200)));
+				itemRepository.save(new Item("4", "Drill Machine", 50, "Units", Double.valueOf(4500)));
+				itemRepository.save(new Item("5", "Wood Plank", 100, "Units", Double.valueOf(800)));
+				itemRepository.save(new Item("6", "Cement Bag", 300, "Units", Double.valueOf(350)));
+				itemRepository.save(new Item("7", "Brick", 5000, "Units", Double.valueOf(5)));
+				itemRepository.save(new Item("8", "Steel Rod", 100, "Units", Double.valueOf(1500)));
+				itemRepository.save(new Item("9", "Sand", 2000, "Kilograms", Double.valueOf(50)));
+				itemRepository.save(new Item("10", "Gravel", 1500, "Kilograms", Double.valueOf(60)));
+				itemRepository.save(new Item("11", "Paint Bucket", 100, "Units", Double.valueOf(900)));
+				itemRepository.save(new Item("12", "Ladder", 50, "Units", Double.valueOf(2500)));
+				itemRepository.save(new Item("13", "Saw", 100, "Units", Double.valueOf(500)));
+				itemRepository.save(new Item("14", "Shovel", 150, "Units", Double.valueOf(400)));
+				itemRepository.save(new Item("15", "Trowel", 200, "Units", Double.valueOf(250)));
+				itemRepository.save(new Item("16", "Plumbing Pipe", 300, "Meters", Double.valueOf(100)));
+				itemRepository.save(new Item("17", "PVC Glue", 200, "Units", Double.valueOf(150)));
+				itemRepository.save(new Item("18", "Tile", 400, "Square Meters", Double.valueOf(300)));
+				itemRepository.save(new Item("19", "Tile Adhesive", 100, "Bags", Double.valueOf(350)));
+				itemRepository.save(new Item("20", "Glass Sheet", 50, "Units", Double.valueOf(2000)));
+				itemRepository.save(new Item("21", "Door Handle", 150, "Units", Double.valueOf(120)));
+				itemRepository.save(new Item("22", "Hinges", 300, "Units", Double.valueOf(60)));
+				itemRepository.save(new Item("23", "Plaster of Paris", 200, "Bags", Double.valueOf(250)));
+				itemRepository.save(new Item("24", "Electrical Wire", 500, "Meters", Double.valueOf(15)));
+				itemRepository.save(new Item("25", "Switchboard", 200, "Units", Double.valueOf(300)));
+				itemRepository.save(new Item("26", "Breaker Panel", 50, "Units", Double.valueOf(4000)));
+				itemRepository.save(new Item("27", "Roof Sheet", 100, "Units", Double.valueOf(2000)));
+				itemRepository.save(new Item("28", "Concrete Mixer", 10, "Units", Double.valueOf(15000)));
+				itemRepository.save(new Item("29", "Safety Helmet", 200, "Units", Double.valueOf(500)));
+				itemRepository.save(new Item("30", "Safety Gloves", 300, "Pairs", Double.valueOf(100)));
+				itemRepository.save(new Item("31", "Welding Rod", 400, "Units", Double.valueOf(75)));
+
 
 			}
 			else {
